Moved bom upload mojo and tests to new packages
