+++ src/main/java/com/pmckeown/mojo/AbstractDependencyTrackMojo.java
@@ -1,10 +1,10 @@
-package com.pmckeown;
+package com.pmckeown.mojo;
 
 import com.pmckeown.rest.client.DependencyTrackClient;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugins.annotations.Parameter;
 
-abstract class AbstractDependencyTrackMojo extends AbstractMojo {
+public abstract class AbstractDependencyTrackMojo extends AbstractMojo {
 
     @Parameter(required = true)
     private String dependencyTrackBaseUrl;
@@ -12,30 +12,30 @@ abstract class AbstractDependencyTrackMojo extends AbstractMojo {
     @Parameter(required = true)
     private String apiKey;
 
-    DependencyTrackClient dependencyTrackClient() {
+    protected DependencyTrackClient dependencyTrackClient() {
         info("Connecting to Dependency Track instance: %s", dependencyTrackBaseUrl);
         return new DependencyTrackClient(dependencyTrackBaseUrl, apiKey);
     }
 
-    void info(String message, Object... params) {
+    protected void info(String message, Object... params) {
         if(getLog().isInfoEnabled()) {
             getLog().info(String.format(message, params));
         }
     }
 
-    void debug(String message, Object... params) {
+    protected void debug(String message, Object... params) {
         if(getLog().isDebugEnabled()) {
             getLog().debug(String.format(message, params));
         }
     }
 
-    void error(String message, Object... params) {
+    protected void error(String message, Object... params) {
         if(getLog().isErrorEnabled()) {
             getLog().error(String.format(message, params));
         }
     }
 
-    void setDependencyTrackBaseUrl(String url) {
+    public void setDependencyTrackBaseUrl(String url) {
         this.dependencyTrackBaseUrl = url;
     }
